enum class fixups - flags patterns.
Fixups in preparation for changing the mojom C++ binding generator to
produce enum classes instead of just enums.  Specifically, address
issues where enums were being as flags and used with bitwise
operations to test for truth.

BUG=
R=jeffbrown@google.com

Review URL: https://codereview.chromium.org/1364453003 .

+++ examples/embedded_app/embedded_app.cc
@@ -87,7 +87,8 @@ class EmbeddedApp
   }
   void OnViewInputEvent(View* view, const EventPtr& event) override {
     if (event->action == EVENT_TYPE_POINTER_UP &&
-        event->flags & EVENT_FLAGS_LEFT_MOUSE_BUTTON) {
+        (static_cast<uint32_t>(event->flags) &
+         static_cast<uint32_t>(EVENT_FLAGS_LEFT_MOUSE_BUTTON))) {
       URLRequestPtr request(URLRequest::New());
       request->url = "http://www.aaronboodman.com/z_dropbox/test.html";
       NavigatorHostPtr navigator_host;
